Move scheduling logic to Api class
Since we need the same initiated SimplyStaticDeployJob
