+++ includes/Api.php
@@ -8,11 +8,12 @@ use WP_REST_Request;
 use WP_REST_Response;
 
 class Api {
+
+    const SCHEDULE_ACTION = 'simply_static_deploy_schedule';
+    const EVENT_BASE = 'simply_static_deploy_event';
+
     const ENDPOINT_MAPPER = [
-        'generate' => 'generate_bundle',
         'generate_single' => 'generate_single',
-        'sync' => 'sync_to_s3',
-        'invalidate' => 'invalidate_cloudfront',
         'simply_static_deploy' => 'simply_static_deploy',
         'poll_status' => 'poll_status',
     ];
@@ -28,6 +29,7 @@ class Api {
 
     public function register() {
         add_action('rest_api_init', [$this, 'register_api_endpoints']);
+        add_action(static::SCHEDULE_ACTION, [$this, 'schedule'], 10, 3);
     }
 
     public function register_api_endpoints() {
@@ -44,6 +46,37 @@ class Api {
                 ]
             );
         }
+    }
+
+    public function schedule(string $time, string $interval): void
+    {
+        $datetime = new DateTime(
+            $time,
+            new DateTimeZone(get_option('timezone_string'))
+        );
+        $event = $this->generate_event_name($datetime, $interval);
+        if (!wp_next_scheduled($event)) {
+            wp_schedule_event($datetime->getTimestamp(), $interval, $event);
+        }
+
+        add_action($event, [$this, 'deploy']);
+    }
+
+    public function deploy(): void
+    {
+        $this->staticDeployJob->start();
+    }
+
+    private function generate_event_name(
+        DateTime $datetime,
+        string $interval
+    ): string {
+        return f\join('_', [
+            static::EVENT_BASE,
+            $interval,
+            $datetime->format('H'),
+            $datetime->format('i'),
+        ]);
     }
 
     public function simply_static_deploy(WP_REST_Request $request) {
